+++ packages/cli-plugin-package/test/midwayfaas3.test.ts
@@ -1,4 +1,4 @@
-import { CommandCore } from '@midwayjs/command-core';
+import { CLIOutputLevel, CommandCore } from '@midwayjs/command-core';
 import { loadSpec } from '@midwayjs/serverless-spec-builder';
 import { PackagePlugin } from '../src/index';
 import { AliyunFCPlugin } from '../../cli-plugin-fc/src/index';
@@ -32,8 +32,50 @@ describe('/test/midwayfaas3.test.ts', () => {
     assert(specFunctions['helloService-httpAllTrigger']);
     assert(specFunctions['helloService-httpTrigger']);
     assert(specFunctions['helloService-ossTrigger']);
-    assert(specFunctions['helloService-coverConfig']);
+    assert(specFunctions['cover-config']);
     assert(specFunctions['helloService-hsfTrigger']);
+  });
+  it('package faas3 trigger', async () => {
+    const baseDir = resolve(__dirname, './fixtures/midwayfaas3_trigger');
+    const buildDir = resolve(baseDir, './.serverless');
+    await remove(buildDir);
+    const logs: any = [];
+    const core = new CommandCore({
+      config: {
+        servicePath: baseDir,
+      },
+      commands: ['package'],
+      service: loadSpec(baseDir),
+      provider: 'aliyun',
+      options: {
+        bundle: true,
+        verbose: true,
+      },
+      log: {
+        log: (...args) => {
+          logs.push(...args);
+        },
+      },
+      outputLevel: CLIOutputLevel.Info,
+    });
+
+    core.addPlugin(PackagePlugin);
+    core.addPlugin(AliyunFCPlugin);
+    await core.ready();
+    await core.invoke(['package']);
+    const specFunctions = (core as any).coreInstance.service.functions;
+    console.log('specFunctions', specFunctions);
+    // TODO: midway bug
+    // assert(specFunctions['aaa1']);
+    assert(
+      !logs.find(log => typeof log === 'string' && log.startsWith('[Verbose]'))
+    );
+    assert(specFunctions['localHttpTest-abc']);
+    assert(specFunctions['localHttpTest-def']);
+    assert(specFunctions['aaa1']);
+    assert(specFunctions['localTest-hello2']);
+    assert(specFunctions['aaa3']);
+    assert(specFunctions['aaa4']);
   });
   it('package cover dep', async () => {
     const baseDir = resolve(__dirname, './fixtures/cover-dep');
@@ -59,7 +101,7 @@ describe('/test/midwayfaas3.test.ts', () => {
     assert(specFunctions['helloService-httpAllTrigger']);
     assert(specFunctions['helloService-httpTrigger']);
     assert(specFunctions['helloService-ossTrigger']);
-    assert(specFunctions['helloService-coverConfig']);
+    assert(specFunctions['cover-config']);
     assert(specFunctions['helloService-hsfTrigger']);
     const pkgJson = JSON.parse(
       readFileSync(join(buildDir, 'package.json'), 'utf-8')